base-kv/base-kv-store-balance-controller/src/main/java/com/baidu/bifromq/basekv/balance/impl/RangeSplitBalancer.java
/*
 * Copyright (c) 2023. The BifroMQ Authors. All Rights Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *    http://www.apache.org/licenses/LICENSE-2.0
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and limitations under the License.
 */

package com.baidu.bifromq.basekv.balance.impl;

import com.baidu.bifromq.basekv.proto.Boundary;
import com.baidu.bifromq.basekv.proto.KVRangeDescriptor;
import com.baidu.bifromq.basekv.proto.KVRangeStoreDescriptor;
import com.baidu.bifromq.basekv.proto.SplitHint;
import com.baidu.bifromq.basekv.raft.proto.ClusterConfig;
import com.baidu.bifromq.basekv.utils.KeySpaceDAG;
import com.google.protobuf.Struct;
import com.google.protobuf.Value;
import java.util.HashMap;
import java.util.Map;
import java.util.NavigableMap;
import java.util.Optional;

public class RangeSplitBalancer extends RuleBasedPlacementBalancer {
    public static final String LOAD_RULE_CPU_USAGE_LIMIT = "maxCpuUsagePerRange";
    public static final String LOAD_RULE_MAX_IO_DENSITY_PER_RANGE = "maxIODensityPerRange";
    public static final String LOAD_RULE_IO_NANOS_LIMIT_PER_RANGE = "ioNanosLimitPerRange";
    public static final String LOAD_RULE_MAX_RANGES_PER_STORE = "maxRangesPerStore";

    private static final String LOAD_TYPE_IO_DENSITY = "ioDensity";
    private static final String LOAD_TYPE_IO_LATENCY_NANOS = "ioLatencyNanos";
    private static final String LOAD_TYPE_CPU_USAGE = "cpu.usage";

    private final String hintType;

    private final Struct defaultLoadRules;

    /**
     * Constructor of StoreBalancer.
     *
     * @param clusterId    the id of the BaseKV cluster which the store belongs to
     * @param localStoreId the id of the store which the balancer is responsible for
     */
    public RangeSplitBalancer(String clusterId,
                              String localStoreId,
                              String hintType,
                              int maxRangesPerRange,
                              double cpuUsageLimit,
                              int maxIoDensityPerRange,
                              long ioNanoLimitPerRange) {
        super(clusterId, localStoreId);
        this.hintType = hintType;
        this.defaultLoadRules = Struct.newBuilder()
            .putFields(LOAD_RULE_CPU_USAGE_LIMIT, Value.newBuilder().setNumberValue(cpuUsageLimit).build())
            .putFields(LOAD_RULE_MAX_IO_DENSITY_PER_RANGE,
                Value.newBuilder().setNumberValue(maxIoDensityPerRange).build())
            .putFields(LOAD_RULE_IO_NANOS_LIMIT_PER_RANGE,
                Value.newBuilder().setNumberValue(ioNanoLimitPerRange).build())
            .putFields(LOAD_RULE_MAX_RANGES_PER_STORE, Value.newBuilder().setNumberValue(maxRangesPerRange).build())
            .build();
    }

    @Override
    protected Struct defaultLoadRules() {
        return defaultLoadRules;
    }

    @Override
    public boolean validate(Struct loadRules) {
        Value cpuUsageLimit = loadRules.getFieldsMap().get(LOAD_RULE_CPU_USAGE_LIMIT);
        if (cpuUsageLimit == null
            || !cpuUsageLimit.hasNumberValue()
            || cpuUsageLimit.getNumberValue() < 0 || cpuUsageLimit.getNumberValue() > 1) {
            return false;
        }
        Value maxIODensityPerRange = loadRules.getFieldsMap().get(LOAD_RULE_MAX_IO_DENSITY_PER_RANGE);
        if (maxIODensityPerRange == null
            || !maxIODensityPerRange.hasNumberValue()
            || maxIODensityPerRange.getNumberValue() < 0) {
            return false;
        }
        Value maxIONanosPerRange = loadRules.getFieldsMap().get(LOAD_RULE_IO_NANOS_LIMIT_PER_RANGE);
        if (maxIONanosPerRange == null
            || !maxIONanosPerRange.hasNumberValue()
            || maxIONanosPerRange.getNumberValue() < 0) {
            return false;
        }
        Value maxRangesPerStore = loadRules.getFieldsMap().get(LOAD_RULE_MAX_RANGES_PER_STORE);
        return maxRangesPerStore != null
            && maxRangesPerStore.hasNumberValue()
            && maxRangesPerStore.getNumberValue() > 0;
    }

    @Override
    protected Map<Boundary, ClusterConfig> doGenerate(Struct loadRules,
                                                      Map<String, KVRangeStoreDescriptor> landscape,
                                                      NavigableMap<Boundary, KeySpaceDAG.LeaderRange> effectiveRoute) {
        double cpuUsageLimit = loadRules.getFieldsMap().get(LOAD_RULE_CPU_USAGE_LIMIT).getNumberValue();
        double maxRangesPerStore = loadRules.getFieldsMap().get(LOAD_RULE_MAX_RANGES_PER_STORE).getNumberValue();
        double maxIODensityPerRange = loadRules.getFieldsMap().get(LOAD_RULE_MAX_IO_DENSITY_PER_RANGE).getNumberValue();
        double ioLatencyLimitPerRange =
            loadRules.getFieldsMap().get(LOAD_RULE_IO_NANOS_LIMIT_PER_RANGE).getNumberValue();
        Map<Boundary, ClusterConfig> expectedRangeLayout = new HashMap<>();
        for (Map.Entry<Boundary, KeySpaceDAG.LeaderRange> entry : effectiveRoute.entrySet()) {
            Boundary boundary = entry.getKey();
            KeySpaceDAG.LeaderRange leaderRange = entry.getValue();
            KVRangeDescriptor rangeDescriptor = leaderRange.descriptor();
            KVRangeStoreDescriptor storeDescriptor = landscape.get(leaderRange.storeId());
            ClusterConfig clusterConfig = rangeDescriptor.getConfig();
            Optional<SplitHint> splitHintOpt = rangeDescriptor
                .getHintsList()
                .stream()
                .filter(h -> h.getType().equals(hintType))
                .findFirst();
            if (splitHintOpt.isPresent()) {
                SplitHint splitHint = splitHintOpt.get();
                double cpuUsage = storeDescriptor.getStatisticsMap().get(LOAD_TYPE_CPU_USAGE);
                double ioDensity = splitHint.getLoadOrDefault(LOAD_TYPE_IO_DENSITY, 0);
                double ioLatencyNanos = splitHint.getLoadOrDefault(LOAD_TYPE_IO_LATENCY_NANOS, 0);
                if (clusterConfig.getNextVotersList().isEmpty() && clusterConfig.getNextLearnersList().isEmpty()
                    && cpuUsage < cpuUsageLimit
                    && ioLatencyNanos < ioLatencyLimitPerRange
                    && ioDensity > maxIODensityPerRange
                    && storeDescriptor.getRangesList().size() < maxRangesPerStore
                    && splitHint.hasSplitKey()) {
                    expectedRangeLayout.put(boundary
                        .toBuilder()
                        .setEndKey(splitHint.getSplitKey())
                        .build(), clusterConfig);
                    expectedRangeLayout.put(boundary
                        .toBuilder()
                        .setStartKey(splitHint.getSplitKey())
                        .build(), clusterConfig);
                } else {
                    expectedRangeLayout.put(boundary, rangeDescriptor.getConfig());
                }
            } else {
                expectedRangeLayout.put(boundary, rangeDescriptor.getConfig());
            }
        }
        return expectedRangeLayout;
    }
}
